Added new landing page

+++ app/aicos_stock_managment/forms.py
@@ -86,7 +86,7 @@ class UmusarurobForm(FlaskForm):
     UmusaruroGrade =  SelectField("Hitamo icyiciro cy'umusaruro", choices=[('good','good'),('normal','normal'), ('bad', 'bad')])
     Quantity =  IntegerField("Ingano y\'umusaruro mu biro (<i><small>kg</small></i>)", validators=[DataRequired()], render_kw={"placeholder": "Injiza ingano y'umusaruro mu biro (kg)"})
     RiceAmount = IntegerField('Igiciro Cy\'umusaruro ku kiro kimwe', validators=[DataRequired()], render_kw={"placeholder": "Injiza igiciro cy\'umusaruro ku kiro kimwe"})
-    UwoAsigaranye =  FloatField("Injiza ibyo asigaranye (<i><small>Byo kurya cg kunywa</small></i>)", validators=[Optional()], render_kw={"placeholder": "Ibiro by'umusaruro asigaranye"})
+    UwoAsigaranye =  IntegerField("Injiza ibyo asigaranye (<i><small>Byo kurya cg kunywa</small></i>)", validators=[Optional()], render_kw={"placeholder": "Ibiro by'umusaruro asigaranye"})
     Gutonoza  = IntegerField('Igiciro cyo gutunganya ibyo asigaranye ku kiro', validators=[Optional()], render_kw={"placeholder": "Injiza igiciro cyo gutunganya ibyo asigaranye ku kiro"})
     submit      =  SubmitField('Emeza')
